Revert "Performance"
