+++ src/database/orders/orders.operations.js
@@ -1,12 +1,10 @@
 import realm from '../init';
-
-import { format, parseISO, isBefore } from 'date-fns';
+import { format, parseISO} from 'date-fns';
 class OrderRealm {
     constructor() {
         this.order = [];
         let firstSyncDate = new Date('November 7, 1973');
         this.lastOrderSync = firstSyncDate;
-        //this.paymentTypeRealm = new PaymentTypeRealm();
     }
 
     truncate() {
@@ -29,8 +27,6 @@ class OrderRealm {
 
         return this.order = formattedArray;
     }
-
-
 
     getOrdersByDate(date) {
         return new Promise(resolve => {
